+++ schema/firebase-config.json
@@ -1221,6 +1221,62 @@
             ],
             "type": "object"
         },
+        "RunSingle": {
+            "additionalProperties": false,
+            "properties": {
+                "baseImageUri": {
+                    "type": "string"
+                },
+                "ignore": {
+                    "items": {
+                        "type": "string"
+                    },
+                    "type": "array"
+                },
+                "output": {
+                    "type": "string"
+                },
+                "postdeploy": {
+                    "anyOf": [
+                        {
+                            "items": {
+                                "type": "string"
+                            },
+                            "type": "array"
+                        },
+                        {
+                            "type": "string"
+                        }
+                    ]
+                },
+                "predeploy": {
+                    "anyOf": [
+                        {
+                            "items": {
+                                "type": "string"
+                            },
+                            "type": "array"
+                        },
+                        {
+                            "type": "string"
+                        }
+                    ]
+                },
+                "region": {
+                    "type": "string"
+                },
+                "serviceId": {
+                    "type": "string"
+                },
+                "source": {
+                    "type": "string"
+                }
+            },
+            "required": [
+                "serviceId"
+            ],
+            "type": "object"
+        },
         "StorageSingle": {
             "additionalProperties": false,
             "properties": {
@@ -2007,6 +2063,19 @@
         },
         "remoteconfig": {
             "$ref": "#/definitions/RemoteConfigConfig"
+        },
+        "run": {
+            "anyOf": [
+                {
+                    "$ref": "#/definitions/RunSingle"
+                },
+                {
+                    "items": {
+                        "$ref": "#/definitions/RunSingle"
+                    },
+                    "type": "array"
+                }
+            ]
         },
         "storage": {
             "anyOf": [
